Extract StepperSequencer logic
This logic is probably the most complex part of this gem, writing it and
refactoring it was much easier with a test!

Extracting out just this part makes is much easier to test without the need for
stubbing/mocking out the collaborators that the Stepper class has.

The testing for this gem is still not very thorough, there is a lot that isn't
tested, but this is a start.